@@ -1,9 +1,10 @@
-const middleware = require('./_common/middleware');
+import dns from 'dns';
+import URL from 'url-parse';
+import middleware from './_common/middleware.js';
 
-const dns = require('dns').promises;
-const URL = require('url-parse');
+// TODO: Fix.
 
-const handler = async (url, event, context) => {
+const mailConfigHandler = async (url, event, context) => {
   try {
     const domain = new URL(url).hostname || new URL(url).pathname;
 
@@ -71,11 +72,11 @@ const handler = async (url, event, context) => {
     } else {
       return {
         statusCode: 500,
-        body: JSON.stringify({ error: error.message }),
+        body: { error: error.message },
       };
     }
   }
 };
 
-module.exports = middleware(handler);
-module.exports.handler = middleware(handler);
+export const handler = middleware(mailConfigHandler);
+export default handler;
