@@ -1,9 +1,9 @@
-const axios = require('axios');
-const middleware = require('./_common/middleware');
+import axios from 'axios';
+import middleware from './_common/middleware.js';
 
 const MOZILLA_TLS_OBSERVATORY_API = 'https://tls-observatory.services.mozilla.com/api/v1';
 
-const handler = async (url) => {
+const tlsHandler = async (url) => {
   try {
     const domain = new URL(url).hostname;
     const scanResponse = await axios.post(`${MOZILLA_TLS_OBSERVATORY_API}/scan?target=${domain}`);
@@ -12,18 +12,18 @@ const handler = async (url) => {
     if (typeof scanId !== 'number') {
       return {
         statusCode: 500,
-        body: JSON.stringify({ error: 'Failed to get scan_id from TLS Observatory' }),
+        body: { error: 'Failed to get scan_id from TLS Observatory' },
       };
     }
     const resultResponse = await axios.get(`${MOZILLA_TLS_OBSERVATORY_API}/results?id=${scanId}`);
     return {
       statusCode: 200,
-      body: JSON.stringify(resultResponse.data),
+      body: resultResponse.data,
     };
   } catch (error) {
     return { error: error.message };
   }
 };
 
-module.exports = middleware(handler);
-module.exports.handler = middleware(handler);
+export const handler = middleware(tlsHandler);
+export default handler;
